+++ janus/src/time.rs
@@ -25,31 +25,3 @@ impl Debug for RealClock {
         write!(f, "{:?}", self.now())
     }
 }
-
-#[doc(hidden)]
-pub mod test_util {
-    use super::Clock;
-    use crate::message::Time;
-    use chrono::{NaiveDate, NaiveDateTime};
-
-    /// A mock clock for use in testing.
-    #[derive(Clone, Copy, Debug)]
-    pub struct MockClock {
-        /// The time that this clock will always return from [`Self::now`]
-        pub(crate) current_time: NaiveDateTime,
-    }
-
-    impl Clock for MockClock {
-        fn now(&self) -> Time {
-            Time::from_naive_date_time(self.current_time)
-        }
-    }
-
-    impl Default for MockClock {
-        fn default() -> Self {
-            Self {
-                current_time: NaiveDate::from_ymd(2001, 9, 9).and_hms(1, 46, 40),
-            }
-        }
-    }
-}
