+++ janus_server/src/lib.rs
@@ -9,5 +9,4 @@ pub mod hpke;
 pub mod message;
 pub mod metrics;
 pub mod task;
-pub mod time;
 pub mod trace;
