+++ monolithic_integration_test/Cargo.toml
@@ -2,11 +2,15 @@
 name = "monolithic_integration_test"
 version = "0.1.0"
 edition = "2021"
+license = "MPL-2.0"
+rust-version = "1.58"
+publish = false
 
 [dev-dependencies]
 chrono = "0.4.19"
 deadpool-postgres = "0.10.1"
 futures = "0.3.21"
+janus = { path = "../janus" }
 janus_server = { path = "../janus_server" }
 lazy_static = "1"
 prio = { git = "https://github.com/divviup/libprio-rs", rev = "229cd9c45924c7dae3ba754a6eb46b2a05ca8451" }  # TODO(brandon): use a numbered version, once a release is cut >0.7.0
