+++ monolithic_integration_test/tests/integration_test.rs
@@ -1,12 +1,14 @@
 use futures::channel::oneshot::Sender;
+use janus::{
+    message::{Duration, Role, TaskId},
+    time::RealClock,
+};
 use janus_server::{
     aggregator::aggregator_server,
     client::{self, Client, ClientParameters},
     datastore::{Crypter, Datastore},
     hpke::test_util::generate_hpke_config_and_private_key,
-    message::{Duration, Role, TaskId},
     task::{AggregatorAuthKey, Task, Vdaf},
-    time::RealClock,
     trace::{install_trace_subscriber, TraceConfiguration},
 };
 use prio::{
